clone bug fixed

+++ lib/zt_admin/clone.rb
@@ -16,6 +16,7 @@
 #   20.06.2021  3.13.0  Bootstrap 5 
 #   13.11.2021  3.15.0  UUID generation
 #   10.01.2022  3.23.0  generic test_helper.rb
+#   20.01.2022  3.25.0  Bug fixed
 ################################################################################
 module ZtAdmin
 
@@ -342,8 +343,10 @@ module ZtAdmin
     end
 
     ### Get generic files in the *test* directory
-    action_report "test/test_helper.rb"
-    FileUtils.cp "#{test}/test_helper.rb", "#{AppRoot}/test/test_helper.rb"
+    if Dir.exists?  "#{AppRoot}/test"
+      action_report "test/test_helper.rb"
+      FileUtils.cp "#{test}/test_helper.rb", "#{AppRoot}/test/test_helper.rb"
+    end
 
     puts colored(MAGENTA, "\n#{TAB}Run commands now (to create db table 'users' & 'samples):")
     puts colored(MAGENTA, "#{TAB*2}rails db:create")
